Change RNG.determine() to use a less-multiply-heavy algo
Multiplication has gotten faster for ulong args on modern processors, but I'd still rather use extra xorshifts than multiplications. Both versions pass PractRand with an increment of 1, when run over 32TB of generated ulongs.  The quality is fine in that case, though not if you use weird other increments or update patterns (which is why we have randomize() too).