Add/implement IReversibleRNG; make Stateful part of IRNG (all can use it)
In the Java version, seeds were always 64-bit longs; in this version, they are strings, and so can be any length (which works great for RNG/Tangle's 127-bit state). While not all RNGs are cleanly reversible, all can have some implementation for the stateful part, even if it returns "" and doesn't change anything in the setter.