The last try was correct only technically; this actually tests the distribution
Some care should be taken when running this test in an automated way. Since some of the RNG objects in rngs are seeded non-deterministically, there's a small chance that those RNGs will all have a much higher or lower average than expected, and trigger a failure on TestInts. The deterministically-seeded RNG objects do fine though.